requirements.txt
torch
torchvision
pillow
pandas
numpy
tqdm
open_clip_torch
scipy
opencv-python
mediapipe
fastapi
gunicorn